Adding support for Prismatic Joints
CCSlideJoint is an Objective-C wrapper (as part of CCBox2D) that adds
support for Prismatic Joints in Box2D.  A Prismatic Joint is a joint
that is fixed along one axis (for example, an object an "rotate" along a
Y axis only while having a fixed X, or vice versa).  This is especially
useful in vehicle objects where you have 2 turning wheels and 2 fixed
wheels - The prismatic joint would be the fixed wheels.