Remove some redundant code.
I wanted to conditionally convert the position, but the conditional logic required is more complex than just converting all the time, even if it's not necessary. Fortunately, the time's it isn't necessary are also the time's when the node's distance from the root is small and the transform calculations should be fast if necessary at all.